Change velocity to double

+++ src/com/MRS/NeckbeardEngine/Enemies/Elite.java
@@ -4,7 +4,7 @@ import com.MRS.NeckbeardEngine.*;
 
 public class Elite extends Enemy {
      
-     public Elite (State state, int x, int y, int xVelocity, int yVelocity, Projectile projectileType, PowerUpPickup heldPowerUp, long timeLine) {
+     public Elite (State state, int x, int y, double xVelocity, double yVelocity, Projectile projectileType, PowerUpPickup heldPowerUp, long timeLine) {
           super(state, x, y, xVelocity, yVelocity, projectileType, heldPowerUp, timeLine);
           health = 3;
      }
